+++ contracts/ClearingHouse.sol
@@ -193,8 +193,8 @@ contract ClearingHouse is
                 })
             );
 
-        // price slippage check
-        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSC");
+        // CH_PSCF: price slippage check fails
+        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSCF");
 
         // if !useTakerBalance, takerBalance won't change, only need to collects fee to oweRealizedPnl
         if (params.useTakerBalance) {
@@ -331,8 +331,8 @@ contract ClearingHouse is
 
         int256 realizedPnl = _settleBalanceAndRealizePnl(trader, params.baseToken, response);
 
-        // price slippage check
-        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSC");
+        // CH_PSCF: price slippage check fails
+        require(response.base >= params.minBase && response.quote >= params.minQuote, "CH_PSCF");
 
         emit LiquidityChanged(
             trader,
@@ -491,8 +491,8 @@ contract ClearingHouse is
         //   trader: here
         //   baseToken: in Exchange.settleFunding()
 
-        // CH_NEO: not empty order
-        require(!IAccountBalance(_accountBalance).hasOrder(trader), "CH_NEO");
+        // CH_NEO: cannot liquidate when there is still order
+        require(!IAccountBalance(_accountBalance).hasOrder(trader), "CH_CLWTISO");
 
         // CH_EAV: enough account value
         require(
@@ -846,8 +846,7 @@ contract ClearingHouse is
     }
 
     function _requireEnoughFreeCollateral(address trader) internal view {
-        // freeCollateral is calculated based on imRatio
-        // CH_NEFCI: not enough account value by imRatio
+        // CH_NEFCI: not enough free collateral by imRatio
         require(
             _getFreeCollateralByRatio(trader, IClearingHouseConfig(_clearingHouseConfig).getImRatio()) >= 0,
             "CH_NEFCI"