+++ test/clearingHouse/ClearingHouse.spec.ts
@@ -120,7 +120,7 @@ describe("ClearingHouse Spec", () => {
         })
 
         it("force error, invalid base token address when setMaxTickCrossedWithinBlock", async () => {
-            await expect(exchange.setMaxTickCrossedWithinBlock(wallet.address, 1)).to.be.revertedWith("EX_ANC")
+            await expect(exchange.setMaxTickCrossedWithinBlock(wallet.address, 1)).to.be.revertedWith("EX_BNC")
         })
 
         it("force error, invalid trustedForwarder address when setTrustedForwarder", async () => {